+++ src/main/java/org/eclipse/keypop/calypso/card/CalypsoCardApiProperties.java
@@ -23,7 +23,7 @@ public final class CalypsoCardApiProperties {
    *
    * @since 1.0.0
    */
-  public static final String VERSION = "2.1";
+  public static final String VERSION = "2.2";
 
   /** Private constructor */
   private CalypsoCardApiProperties() {}