+++ src/main/java/org/eclipse/keypop/calypso/card/transaction/ChannelControl.java
@@ -15,7 +15,9 @@ package org.eclipse.keypop.calypso.card.transaction;
  * Policy for managing the physical channel after a card request is executed.
  *
  * @since 2.0.0
+ * @deprecated This enum is deprecated and will be removed in a future version.
  */
+@Deprecated
 public enum ChannelControl {
 
   /**