+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed JUnit configuration.
 ### Changed
 - Migrated the CI pipeline from Jenkins to GitHub Actions.
+### Deprecated
+- `ChannelControl` enum will be removed in a future version.
+- `TransactionManager.processCommands(ChannelControl channelControl)` method will be removed in a future version.
+### Upgraded
+- API UML specification `2.1.2` -> `2.2.0`.
+- Keypop Reader API `2.0.1` -> `2.1.0`
 
 ## [2.1.2] - 2025-03-20
 ### Added